refactor(kernel): extract provider acquisition barrier
Move dependency planning, staged preflight, bounded credential resolution, and immediate registrar commits from RunBuilder into one cohesive ProviderAcquisition owner. Update durable architecture documentation and the semantic projection.

Verified with a clean independent review, ExDoc warnings-as-errors, focused provider tests, and mix precommit.
